Split Electoral College module into two pieces
One piece (electoral_college.html) is for displaying historical results for elections back to 1964; all vote totals shown in this part are in terms of the number of electors each state had in the year the given election took place.
The other piece (calc.html) is a calculator allowing users to predict the result of the 2012 election; all vote totals shown in this part are in terms of the number of electors each state will have in 2012.

The navigation (lib/map_center/navigation.js) and landing page (index.html) have not been changed pending an overall Electoral College landing page to link to; the changes to lib/map_center/modules/electoral_college.js would also break all existing embeds (as currently written) of the calculator. Because of this, this commit should not be deployed to pbs.org.