Allow for runtime specification of transport
This will allow for simulated clients to either the WebSocket transport
*or* XHR long-polling.

This requires a slight modification of the "socket.io-client" library,
which (although submitted) has not been patched in the upstream
repository. This means the version hosted on NPM is not usable. Instead
of resolving this dependency via NPM, this work includes a patched fork
of the library as a git submodule.